# apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Atom
  version: 0.1.0
  description: |
    Identity and Authorization service for IoT and cloud-native systems.

    All endpoints except `GET /health`, `POST /auth/login`, and
    `GET /.well-known/jwks.json` require a Bearer token:

    ```
    Authorization: Bearer <token>
    ```

    Two token types are accepted:
    - **JWT** (`eyJ...`) — returned by `/auth/login`, short-lived (default 1 h)
    - **API key** (`atom_...`) — long-lived, created per entity

    Endpoints marked **RequireManage** additionally require the caller to hold an
    `allow` + `scope=all` policy binding covering the `manage` capability.
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Reusable components
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT (`eyJ...`) or API key (`atom_<id>_<secret>`)

  # ── Schemas ──────────────────────────────────────────────────────────────
  schemas:
...
    SubjectKind:
      type: string
      enum: [entity, group]
    GrantKind:
      type: string
      enum: [capability, role]
    ScopeKind:
      type: string
      enum: [all, resource_kind, resource]
      description: |
        - `all` — applies to every resource
        - `resource_kind` — applies to all resources of a given kind (`scope_ref` = kind name)
        - `resource` — applies to one specific resource (`scope_ref` = resource UUID)
    Effect:
      type: string
      enum: [allow, deny]
      default: allow

    # Shared error envelope
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: not found

    # ── Auth ───────────────────────────────────────────────────────────────
...
    Resource:
      type: object
      required: [id, kind, attributes, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          example: channel
          description: Free-form kind string — e.g. channel, device, workspace
        name:
          type: string
          nullable: true
        tenant_id:
          type: string
          format: uuid
          nullable: true
        owner_id:
          type: string
          format: uuid
          nullable: true
        attributes:
          type: object
          description: Arbitrary JSONB metadata, usable in ABAC conditions
          example: {"env": "prod", "classification": "confidential"}
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateResource:
      type: object
      required: [kind]
      properties:
        kind:
          type: string
          example: channel
        name:
          type: string
        tenant_id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        attributes:
          type: object
    UpdateResource:
      type: object
      properties:
        name:
          type: string
        attributes:
          type: object
    ResourceList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Resource'
        total:
          type: integer

    # ── Roles ──────────────────────────────────────────────────────────────
    Role:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        tenant_id:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    CreateRole:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: channel-publisher
        tenant_id:
          type: string
          format: uuid
        description:
          type: string
    RoleList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        total:
          type: integer
    AddRoleCapability:
      type: object
      required: [capability_id]
      properties:
        capability_id:
          type: string
          format: uuid

    # ── Capabilities ───────────────────────────────────────────────────────
    Capability:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: publish
        resource_kind:
          type: string
          nullable: true
          description: >
            Scopes this capability to a specific resource kind.
            `null` means the capability applies to all resource kinds.
        description:
          type: string
          nullable: true
    CreateCapability:
      type: object
      required: [name]
      properties:
        name:
          type: string
        resource_kind:
          type: string
          description: Omit to create a capability that applies to all resource kinds
        description:
          type: string
    CapabilityList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Capability'

    # ── Policy bindings ────────────────────────────────────────────────────
    PolicyBinding:
      type: object
      required: [id, subject_kind, subject_id, grant_kind, grant_id, scope_kind, effect, conditions, created_at]
      properties:
        id:
          type: string
          format: uuid
        subject_kind:
          $ref: '#/components/schemas/SubjectKind'
        subject_id:
          type: string
          format: uuid
        grant_kind:
          $ref: '#/components/schemas/GrantKind'
        grant_id:
          type: string
          format: uuid
        scope_kind:
          $ref: '#/components/schemas/ScopeKind'
        scope_ref:
          type: string
          nullable: true
          description: Kind name or resource UUID depending on `scope_kind`
        effect:
          $ref: '#/components/schemas/Effect'
        conditions:
          type: object
          description: >
            Flat dot-path ABAC conditions. All entries must match (AND logic).
            Keys: `entity.attributes.<k>`, `resource.attributes.<k>`, `context.<k>`.
            Empty object `{}` means no conditions.
          example: {"resource.attributes.env": "prod", "context.ip_trusted": "true"}
        created_at:
          type: string
          format: date-time
    CreatePolicyBinding:
      type: object
      required: [subject_kind, subject_id, grant_kind, grant_id, scope_kind]
      properties:
        subject_kind:
          $ref: '#/components/schemas/SubjectKind'
        subject_id:
          type: string
          format: uuid
        grant_kind:
          $ref: '#/components/schemas/GrantKind'
        grant_id:
          type: string
          format: uuid
        scope_kind:
          $ref: '#/components/schemas/ScopeKind'
        scope_ref:
          type: string
          description: >
            For `scope_kind=resource`: resource UUID.
            For `scope_kind=resource_kind`: kind name string.
            For `scope_kind=all`: omit.
        effect:
          $ref: '#/components/schemas/Effect'
        conditions:
          type: object
          example: {}
    PolicyList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PolicyBinding'
        total:
          type: integer

    # ── Authorization check ────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: missing Authorization header
    Forbidden:
      description: Caller lacks `manage` scope=all
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not found
    Conflict:
      description: Already exists (unique constraint)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: already exists

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /resources:
    get:
      operationId: listResources
      summary: List resources
      tags: [Resources]
      parameters:
        - name: kind
          in: query
          schema:
            type: string
        - name: tenant_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Resource list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceList'
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createResource
      summary: Register a resource
      tags: [Resources]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResource'
            example:
              kind: channel
              name: telemetry
              attributes:
                env: prod
      responses:
        "201":
          description: Resource created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /resources/{id}:
    get:
      operationId: getResource
      summary: Get a resource
      tags: [Resources]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateResource
      summary: Update a resource
      tags: [Resources]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateResource'
      responses:
        "200":
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteResource
      summary: Delete a resource
      tags: [Resources]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

  # ── Roles ──────────────────────────────────────────────────────────────
  /roles:
    get:
      operationId: listRoles
      summary: List roles
      tags: [Roles]
      parameters:
        - name: tenant_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Role list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleList'
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createRole
      summary: Create a role
      tags: [Roles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRole'
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "409":
          $ref: '#/components/responses/Conflict'
...
  /roles/{id}/capabilities:
    get:
      operationId: getRoleCapabilities
      summary: List capabilities in a role
      tags: [Roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Capability list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityList'
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: addRoleCapability
      summary: Add a capability to a role
      tags: [Roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRoleCapability'
      responses:
        "204":
          description: Added
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
...
  /capabilities:
    get:
      operationId: listCapabilities
      summary: List capabilities
      description: Passing `resource_kind` returns capabilities for that kind plus universal ones (`resource_kind=null`).
      tags: [Capabilities]
      parameters:
        - name: resource_kind
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Capability list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityList'
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createCapability
      summary: Create a capability
      description: "**RequireManage.**"
      tags: [Capabilities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCapability'
      responses:
        "201":
          description: Capability created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capability'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
...
  /policies:
    get:
      operationId: listPolicies
      summary: List policy bindings
      tags: [Policies]
      parameters:
        - name: subject_id
          in: query
          schema:
            type: string
            format: uuid
        - name: subject_kind
          in: query
          schema:
            $ref: '#/components/schemas/SubjectKind'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Policy binding list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyList'
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createPolicy
      summary: Create a policy binding
      description: "**RequireManage.**"
      tags: [Policies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePolicyBinding'
            examples:
              rbac:
                summary: Role grant over a resource kind
                value:
                  subject_kind: entity
                  subject_id: "550e8400-e29b-41d4-a716-446655440000"
                  grant_kind: role
                  grant_id: "7c4b7f1e-0000-0000-0000-000000000001"
                  scope_kind: resource_kind
                  scope_ref: channel
                  effect: allow
              abac:
                summary: Capability with ABAC condition
                value:
                  subject_kind: entity
                  subject_id: "550e8400-e29b-41d4-a716-446655440000"
                  grant_kind: capability
                  grant_id: "7c4b7f1e-0000-0000-0000-000000000002"
                  scope_kind: resource_kind
                  scope_ref: secret
                  effect: allow
                  conditions:
                    "resource.attributes.env": "prod"
              deny:
                summary: Explicit deny for a specific entity
                value:
                  subject_kind: entity
                  subject_id: "550e8400-e29b-41d4-a716-446655440000"
                  grant_kind: role
                  grant_id: "7c4b7f1e-0000-0000-0000-000000000001"
                  scope_kind: all
                  effect: deny
      responses:
        "201":
          description: Policy binding created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyBinding'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
...